// web/src/types/flow.ts
export type Widget<Structure> = {
	id: string
	version: string
	type: string
	structure: Structure
}
// ...
	export type Api = Widget<{
		method: string
		domain: string
		path: string
		uri: string
		url: string
		proxy: string
		request: {
			params_query: Widget<Schema>
			params_path: Widget<Schema>
			body_type: string
			body: Widget<Schema>
			headers: Widget<Schema>
		}
	}>
// ...
export enum ComponentTypes {
	/** 表单组件 */
	Form = "form",
	/** 控件组件 */
	Widget = "widget",
	/** 条件组件 */
	Condition = "condition",
	/** api组件 */
	Api = "api",
	/** value组件 */
	Value = "value",
}